added scripts helpers
